refactor(acp): extract handler application workflows
